feat: Use datafusion planner as the default execution strategy
